refactor: rename agent commands to session commands for clarity
Refactors the command structure to properly separate agent (configuration)
and session (running process) management:

- Move all session management from `agent` to new `session` command
  - `amux session run <agent>` - Run an agent session
  - `amux session list` - List active sessions
  - `amux session attach <session>` - Attach to running session
  - `amux session stop <session>` - Stop a session
  - `amux session logs <session>` - View session output

- Simplify `agent` command to only view configurations
  - `amux agent list` - List configured agents
  - `amux agent show <agent>` - Show agent configuration
  - Configuration changes will be done via `amux config edit`

- Maintain global aliases for backward compatibility
  - `amux run` → `amux session run`
  - `amux ps` → `amux session list`
  - `amux attach` → `amux session attach`
  - `amux tail` → `amux session logs -f`

This change clarifies that agents are static configurations while
sessions are running process instances.

Fixes #100
