+++ internal/cli/commands/agent/agent.go
@@ -1,53 +1,25 @@
-// Package agent implements agent-related CLI commands.
+// Package agent implements agent configuration CLI commands.
 package agent
 
 import (
 	"github.com/spf13/cobra"
 )
 
-var (
-	// Flags for agent run
-	runWorkspace     string
-	runCommand       string
-	runEnv           []string
-	runInitialPrompt string
-
-	// Flags for agent logs
-	followLogs bool
-)
-
 // Command returns the agent command
 func Command() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "agent",
-		Short: "Manage AI agent sessions",
-		Long: `Manage AI agent sessions in multiplexed workspaces.
+		Short: "View agent configurations",
+		Long: `View configured AI agents.
 
-Run multiple AI agents concurrently in isolated workspaces,
-attach to running sessions, and manage agent lifecycle.`,
+Agents are static configurations that define which AI tool to run,
+default commands, and environment variables. To modify agent configurations,
+use 'amux config edit'.`,
 	}
-
-	// Initialize config command
-	initAgentConfig()
 
 	// Add subcommands
-	cmd.AddCommand(runCmd())
 	cmd.AddCommand(listCmd())
-	cmd.AddCommand(attachCmd())
-	cmd.AddCommand(stopCmd())
-	cmd.AddCommand(logsCmd())
-	cmd.AddCommand(agentConfigCmd)
+	cmd.AddCommand(showCmd())
 
 	return cmd
-}
-
-// TailCommand returns the tail command (alias for agent logs -f)
-func TailCommand() *cobra.Command {
-	return &cobra.Command{
-		Use:   "tail <session>",
-		Short: "Follow agent session logs in real-time",
-		Long:  "Continuously stream output from an agent session. Similar to 'agent logs -f'.",
-		Args:  cobra.ExactArgs(1),
-		RunE:  tailAgentLogs,
-	}
 }
