+++ internal/cli/commands/session/helpers_test.go
@@ -1,4 +1,4 @@
-package agent
+package session
 
 import (
 	"fmt"
