+++ internal/cli/commands/session/logs_test.go
@@ -1,4 +1,4 @@
-package agent
+package session
 
 import (
 	"testing"
@@ -6,17 +6,17 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
-func TestTailAgentLogsFlag(t *testing.T) {
+func TestTailSessionLogsFlag(t *testing.T) {
 	// Store original value
 	originalFollow := followLogs
 	defer func() { followLogs = originalFollow }()
 
-	// Test that tailAgentLogs function sets the follow flag
+	// Test that tailSessionLogs function sets the follow flag
 	// This is a simple unit test to verify the flag behavior
 	followLogs = false
 	assert.False(t, followLogs)
 
-	// The tailAgentLogs function sets followLogs = true
+	// The tailSessionLogs function sets followLogs = true
 	// We test this by checking the implementation pattern
-	// (The actual function would call viewAgentLogs which requires a full session setup)
+	// (The actual function would call viewSessionLogs which requires a full session setup)
 }
